separate classes and structs -wip
