+++ SharedProject1/Configurations/ClippyAnimation.cs
@@ -50,132 +50,6 @@
         Writing
     }
 
-    public enum MerlinAnimations
-    {
-        MoveLeft,
-        Congratulate,
-        Hide,
-        Pleased,
-        Acknowledge,
-        Thinking,
-        Suggest,
-        Explain,
-        Decline,
-        DontRecognize,
-        Writing,
-        Write,
-        Idle3_2,
-        Idle3_1,
-        Congratulate2,
-        StartListening,
-        Idle2_2,
-        Announce,
-        GetAttention,
-        Idle2_1,
-        GestureLeft,
-        Surprised,
-        GestureRight,
-        Idle1_4,
-        LookLeftReturn,
-        GestureUp,
-        Idle1_1,
-        Idle1_3,
-        Idle1_2,
-        Read,
-        Processing,
-        Wave,
-        DoMagic1,
-        DoMagic2,
-        LookRight,
-        Alert,
-        MoveRight,
-        Reading,
-        GetAttentionContinued,
-        WriteContinued,
-        Confused,
-        LookRightBlink,
-        Search,
-        Uncertain,
-        LookLeft,
-        LookDownReturn,
-        Hearing4,
-        LookUpReturn,
-        Hearing1,
-        Greet,
-        Hearing3,
-        WriteReturn,
-        Hearing2,
-        GetAttentionReturn,
-        RestPose,
-        LookDownBlink,
-        LookUpBlink,
-        Think,
-        Blink,
-        Show,
-        LookRightReturn,
-        StopListening,
-        MoveDown,
-        ReadContinued,
-        LookDown,
-        Sad,
-        Process,
-        LookUp,
-        GestureDown,
-        ReadReturn,
-        Searching,
-        MoveUp,
-        LookLeftBlink
-    }
-
-    public enum GeniusAnimations
-    {
-        DeepIdle1,
-        Congratulate,
-        Hide,
-        SendMail,
-        Thinking,
-        Explain,
-        Save,
-        Print,
-        LookRight,
-        GetAttention,
-        Alert,
-        GetTechy,
-        GestureUp,
-        Idle11,
-        Processing,
-        LookUpRight,
-        Idle7,
-        Idle6,
-        Idle5,
-        Idle4,
-        Idle3,
-        Idle2,
-        Idle1,
-        Idle0,
-        Idle9,
-        Idle8,
-        LookLeft,
-        LookUpLeft,
-        CheckingSomething,
-        Hearing1,
-        GetWizardy,
-        GestureLeft,
-        Wave,
-        Goodbye,
-        GestureRight,
-        Writing,
-        LookDownRight,
-        GetArtsy,
-        Show,
-        LookDown,
-        Searching,
-        EmptyTrash,
-        Greeting,
-        LookUp,
-        GestureDown,
-        RestPose,
-        LookDownLeft
-
-    }
+    
+   
 }
