+++ SharedProject1/Configurations/GeniusAnimations.cs
@@ -1,0 +1,55 @@
+﻿
+namespace SharedProject1.Configurations
+{
+    public enum GeniusAnimations
+    {
+        DeepIdle1,
+        Congratulate,
+        Hide,
+        SendMail,
+        Thinking,
+        Explain,
+        Save,
+        Print,
+        LookRight,
+        GetAttention,
+        Alert,
+        GetTechy,
+        GestureUp,
+        Idle11,
+        Processing,
+        LookUpRight,
+        Idle7,
+        Idle6,
+        Idle5,
+        Idle4,
+        Idle3,
+        Idle2,
+        Idle1,
+        Idle0,
+        Idle9,
+        Idle8,
+        LookLeft,
+        LookUpLeft,
+        CheckingSomething,
+        Hearing1,
+        GetWizardy,
+        GestureLeft,
+        Wave,
+        Goodbye,
+        GestureRight,
+        Writing,
+        LookDownRight,
+        GetArtsy,
+        Show,
+        LookDown,
+        Searching,
+        EmptyTrash,
+        Greeting,
+        LookUp,
+        GestureDown,
+        RestPose,
+        LookDownLeft
+
+    }
+}
