+++ SharedProject1/Configurations/MerlinAnimations.cs
@@ -1,0 +1,84 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SharedProject1.Configurations
+{
+    public enum MerlinAnimations
+    {
+        MoveLeft,
+        Congratulate,
+        Hide,
+        Pleased,
+        Acknowledge,
+        Thinking,
+        Suggest,
+        Explain,
+        Decline,
+        DontRecognize,
+        Writing,
+        Write,
+        Idle3_2,
+        Idle3_1,
+        Congratulate2,
+        StartListening,
+        Idle2_2,
+        Announce,
+        GetAttention,
+        Idle2_1,
+        GestureLeft,
+        Surprised,
+        GestureRight,
+        Idle1_4,
+        LookLeftReturn,
+        GestureUp,
+        Idle1_1,
+        Idle1_3,
+        Idle1_2,
+        Read,
+        Processing,
+        Wave,
+        DoMagic1,
+        DoMagic2,
+        LookRight,
+        Alert,
+        MoveRight,
+        Reading,
+        GetAttentionContinued,
+        WriteContinued,
+        Confused,
+        LookRightBlink,
+        Search,
+        Uncertain,
+        LookLeft,
+        LookDownReturn,
+        Hearing4,
+        LookUpReturn,
+        Hearing1,
+        Greet,
+        Hearing3,
+        WriteReturn,
+        Hearing2,
+        GetAttentionReturn,
+        RestPose,
+        LookDownBlink,
+        LookUpBlink,
+        Think,
+        Blink,
+        Show,
+        LookRightReturn,
+        StopListening,
+        MoveDown,
+        ReadContinued,
+        LookDown,
+        Sad,
+        Process,
+        LookUp,
+        GestureDown,
+        ReadReturn,
+        Searching,
+        MoveUp,
+        LookLeftBlink
+    }
+
+}
