+++ SharedProject1/SharedProject1.projitems
@@ -26,7 +26,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)Configurations\ClippyAnimation.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Configurations\ColumnRow.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Configurations\Frame.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Configurations\GeniusAnimations.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Configurations\Legacy\LegacySingleAnimation.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Configurations\MerlinAnimations.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Constants.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)GlobalSuppressions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IClippyVSSettings.cs" />
